+++ crates/fluxheim-server/src/lib.rs
@@ -39,6 +39,7 @@ mod native_http1_route_proxy;
 mod native_http1_route_redirect;
 mod native_http1_route_request_headers;
 mod native_http1_route_response_headers;
+mod native_http1_route_rewrite;
 mod native_http1_static_web;
 #[cfg(any(feature = "tls-rustls-backend", feature = "tls-openssl-backend"))]
 mod native_http1_tls;